Remove `Device` and `Device-type` from base image detail messages
In our build process, we create groups of all identical base images and for each group, just one image is built while others are created by doing `docker tag` so the `Device` and `Device-type` on the tagged images are not correct.

Users already know the device name from the base image name so these information are not really important and we should remove them from the messages, otherwise it can be confusing for users.

Change-type: patch
Signed-off-by: Trong Nghia Nguyen <nghiant2710@gmail.com>